requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.27
yolov5==7.0.11
pandas==1.3.3
